Правки по ревью https://online.sbis.ru/opendoc.html?guid=0f34fd28-797b-4ec2-b0f6-a4dcd28a96bc
