--- sabycom/src/main/res/drawable/sabycom_background.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-
-    <solid android:color="#ffffff" />
-
-    <corners
-        android:topLeftRadius="18dp"
-        android:topRightRadius="18dp" />
-
-</shape>
