+++ sabycom/src/main/res/layout/sabycom_dialog.xml
@@ -1,11 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/web_view_container"
+<WebView xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/web_view"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <WebView
-        android:id="@+id/web_view"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" />
-</FrameLayout>
+    android:layout_height="match_parent" />
