+++ sabycomdemo/src/main/res/layout/activity_sabycom_demo.xml
@@ -9,23 +9,23 @@
     <TextView
         android:id="@+id/counter"
         android:layout_width="wrap_content"
-        android:layout_height="20dp"
-        android:layout_marginBottom="-15dp"
+        android:layout_height="wrap_content"
         android:background="@drawable/counter_background"
         android:textAlignment="center"
+        android:translationZ="3dp"
+        android:layout_marginEnd="4dp"
+        android:layout_marginTop="4dp"
         android:textColor="@color/design_default_color_surface"
-        android:translationZ="10dp"
-        android:paddingStart="4dp"
-        android:paddingEnd="4dp"
         app:layout_constraintBottom_toTopOf="@+id/show_sabycom"
         app:layout_constraintEnd_toEndOf="@+id/show_sabycom"
+        app:layout_constraintTop_toTopOf="@+id/show_sabycom"
         tools:text="1234" />
 
     <Button
         android:id="@+id/show_sabycom"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Sabycom"
+        android:text="@string/sabycom_button_text"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
